Модифицированы старые юнит-тесты, написанные в 1 ЛР

+++ UnitTestProject/UnitTests.cs
@@ -8,13 +8,14 @@ namespace UnitTestProject
     [TestClass]
     public class UnitTestsInventoryForm
     {
+        private InventoryForm form; // добавлена ссылка на объект формы
+
         [TestInitialize]
         public void Setup()
         {
-            if (File.Exists("inventory.txt"))
-            {
-                File.Delete("inventory.txt");
-            }
+            form = new InventoryForm(); // каждый тест - создание объекта.
+            // Теперь можно не прописывать создание формы в каждом новом тесте
+            if (File.Exists("inventory.txt")) { File.Delete("inventory.txt"); }
         }
 
         /*
@@ -27,7 +28,6 @@ namespace UnitTestProject
         [TestMethod]
         public void AddItem_FullFields_StrokeAdd()
         {
-            var form = new InventoryForm();
             form.AddItem("Футбольный мяч", "20", "3999", "Футбол");
             Assert.AreEqual(1, form.ReturnItemsCount(), "Строчка с заполненными полями НЕ добавлена в список!");
         }
@@ -35,7 +35,6 @@ namespace UnitTestProject
         [TestMethod]
         public void AddItem_EmptyFields_Error()
         {
-            var form = new InventoryForm();
             form.AddItem("Хоккейная клюшка", "", "", "");
             Assert.AreEqual(0, form.ReturnItemsCount(), "Строчка с пустыми полями добавлена в список!");
         }
@@ -43,7 +42,6 @@ namespace UnitTestProject
         [TestMethod]
         public void AddItem_WrongQuantity_Error ()
         {
-            var form = new InventoryForm();
             form.AddItem("Баскетбольное кольцо", "quantity", "4200", "Баскетбол");
             Assert.AreEqual(0, form.ReturnItemsCount(), "Строчка с неверным кол-вом попала в список!");
         }
@@ -51,10 +49,13 @@ namespace UnitTestProject
         [TestMethod]
         public void AddItem_WrongPrice_Error()
         {
-            var form = new InventoryForm();
             form.AddItem("Боксерский шлем", "10", "price", "Бокс");
             Assert.AreEqual(0, form.ReturnItemsCount(), "Строчка с неверной ценой попала в список!");
         }
+
+
+
+
     }
 
     [TestClass]
@@ -63,10 +64,7 @@ namespace UnitTestProject
         [TestInitialize]
         public void Setup ()
         {
-            if (File.Exists("inventory.txt"))
-            {
-                File.Delete("inventory.txt");
-            }
+            if (File.Exists("inventory.txt")) { File.Delete("inventory.txt"); }
         }
 
         [TestMethod]
@@ -120,12 +118,14 @@ namespace UnitTestProject
             Console.WriteLine("Корректность записи строки: " + result);
             Assert.IsFalse(result);
         }
-
-
-
-
-
     }
+
+
+
+
+
+
+
